Apply nonces to CSP in central module
